+++ src/quotes/nicknameParts.js
@@ -1,0 +1,38 @@
+// nickname generation parts
+export const ADVERBS = [
+  'neatly',
+  'uncleanly',
+  'bluntly',
+  'trivially',
+  'bravely',
+  'spiritually',
+  'so',
+  'lazily',
+  'ruthlessly',
+  'eagerly'
+];
+
+export const ADJECTIVES = [
+  'abusive',
+  'leaking',
+  'reverting',
+  'raging',
+  'ample',
+  'admirable',
+  'arctic',
+  'corny',
+  'dense',
+  'feisty'
+];
+export const NOUNS = [
+  'boxer',
+  'walrus',
+  'trickster',
+  'typist',
+  'kidoh',
+  'soul',
+  'minion',
+  'telegraph',
+  'painter',
+  'jungleboy'
+];
